Moved convert method to helpers
